Add option to find the shortest macro (#69)
* Add shortest_duration option (solvers)
* Prioritize step count over macro duration
* Add shortest_duration option (ui)
* Change backtrack id type to usize
* Change level type to u8 & Fix clippy warnings

+++ simulator/src/actions.rs
@@ -46,7 +46,7 @@ pub enum ComboAction {
 }
 
 impl Action {
-    pub const fn level_requirement(self) -> u32 {
+    pub const fn level_requirement(self) -> u8 {
         match self {
             Action::BasicSynthesis => 1,
             Action::BasicTouch => 5,
